# backend/api-contracts/src/main/java/com/ticketplatform/api/enums/ConfidenceScore.java
package com.ticketplatform.api.enums;

/**
 * AI suggestion confidence level (0-100 scale).
 */
public enum ConfidenceScore {
    /** 0-25 */
    LOW,
    /** 26-50 */
    MEDIUM,
    /** 51-75 */
    HIGH,
    /** 76-100 */
    VERY_HIGH
}